chore: Update template values for new project
Follow the TODOs to update the values provided by the template, several
steps haven't been followed yet so have been left as TODOs or in the
README as a TODO section.

TISNEW-4785
