+++ .aws/task-definition.json
@@ -1,8 +1,8 @@
 {
   "containerDefinitions": [
     {
-      "name": "tis-template",
-      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-template:1",
+      "name": "tis-revalidation-integration",
+      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-revalidation-integration:1",
       "portMappings": [
         {
           "containerPort": 8080
@@ -11,7 +11,7 @@
       "environment": [
         {
           "name": "TITLE",
-          "value": "tis-template"
+          "value": "tis-revalidation-integration"
         },
         {
           "name": "AWS_REGION",
@@ -21,9 +21,9 @@
       "logConfiguration": {
         "logDriver": "awslogs",
         "options": {
-          "awslogs-group": "awslogs-tis-template",
+          "awslogs-group": "awslogs-tis-revalidation-integration",
           "awslogs-region": "eu-west-2",
-          "awslogs-stream-prefix": "awslogs-tis-template"
+          "awslogs-stream-prefix": "awslogs-tis-revalidation-integration"
         }
       },
       "secrets": [
@@ -47,7 +47,7 @@
     }
   ],
   "executionRoleArn": "ecsTaskExecutionRole",
-  "family": "tis-template",
+  "family": "tis-revalidation-integration",
   "requiresCompatibilities": [
     "FARGATE"
   ],
