+++ settings.gradle
@@ -1,2 +1,1 @@
-// TODO: Update project name.
-rootProject.name = 'template'
+rootProject.name = 'tis-revalidation-integration'
