+++ src/main/java/uk/nhs/hee/tis/revalidation/integration/RevalidationIntegrationApplication.java
@@ -19,25 +19,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package uk.nhs.hee.tis.template.config;
+package uk.nhs.hee.tis.revalidation.integration;
 
-import io.sentry.spring.SentryExceptionResolver;
-import io.sentry.spring.SentryServletContextInitializer;
-import org.springframework.boot.web.servlet.ServletContextInitializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.HandlerExceptionResolver;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@Configuration
-public class SentryConfiguration {
+@SpringBootApplication
+public class RevalidationIntegrationApplication {
 
-  @Bean
-  public HandlerExceptionResolver sentryExceptionResolver() {
-    return new SentryExceptionResolver();
-  }
-
-  @Bean
-  public ServletContextInitializer sentryServletContextInitilizer() {
-    return new SentryServletContextInitializer();
+  public static void main(String[] args) {
+    SpringApplication.run(RevalidationIntegrationApplication.class);
   }
 }
