+++ src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  # TODO: Update server port and context path to a unique value.
+  # TODO: Update server port to a unique value.
   port: 8080
   servlet:
-    context-path: /template
+    context-path: /revalidation
